% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_apply.R
\name{apply_fun}
\alias{apply_fun}
\title{apply_fun}
\usage{
apply_fun(data, coeff, eq_list)
}
\arguments{
\item{data}{data on which predictions are going to be calculated}

\item{coeff}{list of equation's coefficients}

\item{eq_list}{list of equations to apply}
}
\description{
Function to make predictions based on inputed equations
}
\examples{

eq_t <- find_fun(c("Ta", "H", "R"), eq_list)
}
