% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFun.R
\name{best_tr}
\alias{best_tr}
\title{Best Model Function}
\usage{
best_tr(y, x, type = "all", return = "all", cn = FALSE)
}
\arguments{
\item{y}{dependent variable}

\item{x}{regressor}

\item{type}{type of model to use: linear - "lm", exponential - "exp" and power trend "power". Default "all" use all three types.}

\item{return}{what kind of data to return:
"data" - fitted values;
"plot" - ggplot of fitted values;
"table" - contains information about coefficint of determination (R2), standard deviation (sd), mean absolute error (mae),
mean relative error (mre) and Kling-Gupta efficiency (KGE);
"all" - returns all;}

\item{cn}{whether to keep column names set by function (default is FALSE)}

\item{data}{data to compute trends}
}
\description{
This function estimates coefficients for three different types of model.
}
\examples{
est <- best_tr(dta$evaporation, dta$radiation, cn=TRUE)
df <- as.data.frame(est$dta.n)
est$tab
plot_grid(plotlist = est$gg.all)

}
