% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_coef_table.R
\name{compute_coef_table}
\alias{compute_coef_table}
\title{compute_coef_table}
\usage{
compute_coef_table(rad)
}
\arguments{
\item{rad}{input data.frame}
}
\description{
Function to compute possible combinations of equations and their coefficients, returns parameters in three forms:
list of equations with every coefficient (all_coef);
table (ult_tab), where columns are variables and rows are equations, contains information about KGE and MRE with final equation for every row (for each combination);
table (koefs), which contains coefficients and equation for each combination.
}
\examples{
recomended
options(stringsAsFactors = FALSE)
}
