% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFun.R
\name{extractSUM}
\alias{extractSUM}
\title{extractSUM}
\usage{
extractSUM(x, y)
}
\arguments{
\item{x}{first equation}

\item{y}{second equation}
}
\description{
Function to calculate sum of extracted constant terms from equations
}
\examples{
extractSUM("E = 0.4254R + 1.0341", "E = 7.5687Tvzd - 82.29")
-81.2559
}
