% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFun.R
\name{extractVAR}
\alias{extractVAR}
\title{extractVAR}
\usage{
extractVAR(e)
}
\arguments{
\item{e}{equation}
}
\description{
Function to extract variables from equations.
}
\examples{
extractVAR("E = 0.0156R^1.0156 - 0.8136H + 0.0106Tw + 0.4647")
"R"  "H"  "Tw"
}
