% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_apply.R
\name{find_fun}
\alias{find_fun}
\title{find_fun}
\usage{
find_fun(IN, E)
}
\arguments{
\item{IN}{string or vector of variables}

\item{E}{list of equations to search in}
}
\description{
Function to search equations with selected variables
}
\examples{

eq_t <- find_fun(c("Ta", "H", "R"), eq_list)
"E = 0.0169R - 1.0099H + 0.022Ta + 0.5102" "E = 0.0169R + 0.0374Ta - 0.4965"          "E = 0.0156R^1.0156 + 0.0336Ta - 0.4588"
"E = 0.0156R^1.0156 - 0.8136H + 0.6596"    "E = 0.0169R - 1.0099H + 0.829"
}
