% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFun.R
\name{numextract}
\alias{numextract}
\title{numextract}
\usage{
numextract(string, ret = "s")
}
\arguments{
\item{string}{equation}

\item{ret}{whether returned value is going to be string ("s", default) or numeric ("n")}
}
\description{
Function to extract constant term from the equation string for further manipulations
}
\examples{
numextract("E = 0.4254*R + 1.0341", "n")
1.0341
}
