% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFun.R
\name{scatter_fun}
\alias{scatter_fun}
\title{scatter_fun}
\usage{
scatter_fun(y, x, ysim, title = "Titulek")
}
\arguments{
\item{y}{dependent variable}

\item{x}{regressors}

\item{ysim}{fitted values}

\item{title}{plot title}
}
\description{
This function plots trend.
}
\examples{
scatter_fun()
}
